+++ src/Microsoft.DotNet.Interactive.AIUtilities/ScoredItem.cs
@@ -1,0 +1,7 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+// ReSharper disable once CheckNamespace
+namespace System.Collections;
+
+public record ScoredItem<T>(T Item, float Score);
